+++ RELEASES.md
@@ -8,8 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 vNext
 -----
 
+### aoms
+
+- Improved: Enhanced errors with stack trace and custom descriptions
+
 ### aomc
 
+- Improved: Enhanced errors with stack trace and custom descriptions
 - Changed: Removed parameter "system" from method GetClasses()
 - Fixed: Empty list of classes is returned even if an error occurs during unmarshalling the JSON body from the response